+++ lib/prism/translation.rb
@@ -7,11 +7,11 @@ module Prism
   module Translation # steep:ignore
     autoload :Parser, "prism/translation/parser"
     autoload :ParserCurrent, "prism/translation/parser_current"
-    autoload :Parser33, "prism/translation/parser33"
-    autoload :Parser34, "prism/translation/parser34"
-    autoload :Parser35, "prism/translation/parser35"
-    autoload :Parser40, "prism/translation/parser40"
-    autoload :Parser41, "prism/translation/parser41"
+    autoload :Parser33, "prism/translation/parser_versions"
+    autoload :Parser34, "prism/translation/parser_versions"
+    autoload :Parser35, "prism/translation/parser_versions"
+    autoload :Parser40, "prism/translation/parser_versions"
+    autoload :Parser41, "prism/translation/parser_versions"
     autoload :Ripper, "prism/translation/ripper"
     autoload :RubyParser, "prism/translation/ruby_parser"
   end
