--- lib/prism/translation/parser33.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-# :markup: markdown
-
-module Prism
-  module Translation
-    # This class is the entry-point for Ruby 3.3 of `Prism::Translation::Parser`.
-    class Parser33 < Parser
-      def version # :nodoc:
-        33
-      end
-    end
-  end
-end
