--- lib/prism/translation/parser34.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-# :markup: markdown
-
-module Prism
-  module Translation
-    # This class is the entry-point for Ruby 3.4 of `Prism::Translation::Parser`.
-    class Parser34 < Parser
-      def version # :nodoc:
-        34
-      end
-    end
-  end
-end
