--- lib/prism/translation/parser35.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-# :markup: markdown
-
-module Prism
-  module Translation
-    Parser35 = Parser40 # :nodoc:
-  end
-end
