--- lib/prism/translation/parser41.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-# :markup: markdown
-
-module Prism
-  module Translation
-    # This class is the entry-point for Ruby 4.1 of `Prism::Translation::Parser`.
-    class Parser41 < Parser
-      def version # :nodoc:
-        41
-      end
-    end
-  end
-end
