+++ prism.gemspec
@@ -99,11 +99,7 @@ Gem::Specification.new do |spec|
     "lib/prism/translation.rb",
     "lib/prism/translation/parser.rb",
     "lib/prism/translation/parser_current.rb",
-    "lib/prism/translation/parser33.rb",
-    "lib/prism/translation/parser34.rb",
-    "lib/prism/translation/parser35.rb",
-    "lib/prism/translation/parser40.rb",
-    "lib/prism/translation/parser41.rb",
+    "lib/prism/translation/parser_versions.rb",
     "lib/prism/translation/parser/builder.rb",
     "lib/prism/translation/parser/compiler.rb",
     "lib/prism/translation/parser/lexer.rb",
@@ -123,11 +119,7 @@ Gem::Specification.new do |spec|
     "rbi/prism/reflection.rbi",
     "rbi/prism/string_query.rbi",
     "rbi/prism/translation/parser.rbi",
-    "rbi/prism/translation/parser33.rbi",
-    "rbi/prism/translation/parser34.rbi",
-    "rbi/prism/translation/parser35.rbi",
-    "rbi/prism/translation/parser40.rbi",
-    "rbi/prism/translation/parser41.rbi",
+    "rbi/prism/translation/parser_versions.rbi",
     "rbi/prism/translation/ripper.rbi",
     "rbi/prism/visitor.rbi",
     "sig/prism.rbs",
