--- rbi/prism/translation/parser33.rbi
@@ -1,6 +1,0 @@
-# typed: strict
-
-class Prism::Translation::Parser33 < Prism::Translation::Parser
-  sig { override.returns(Integer) }
-  def version; end
-end
