--- rbi/prism/translation/parser35.rbi
@@ -1,3 +1,0 @@
-# typed: strict
-
-Prism::Translation::Parser35 = Prism::Translation::Parser40
