+++ rbi/prism/translation/parser_versions.rbi
@@ -1,0 +1,23 @@
+# typed: strict
+
+class Prism::Translation::Parser33 < Prism::Translation::Parser
+  sig { override.returns(Integer) }
+  def version; end
+end
+
+class Prism::Translation::Parser34 < Prism::Translation::Parser
+  sig { override.returns(Integer) }
+  def version; end
+end
+
+class Prism::Translation::Parser40 < Prism::Translation::Parser
+  sig { override.returns(Integer) }
+  def version; end
+end
+
+Prism::Translation::Parser35 = Prism::Translation::Parser40
+
+class Prism::Translation::Parser40 < Prism::Translation::Parser
+  sig { override.returns(Integer) }
+  def version; end
+end
